requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pdfminer.six==20250506
phonenumbers==9.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
python-multipart==0.0.20
python-stdnum==2.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
